# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Locadora API",
        "description": "API para sistema de locadora de veículos",
        "contact": {
            "email": "admin@locadora.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/rentals/checkout": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Realizar check-out (Iniciar locação)",
                "operationId": "12c39ccfd72dc64239ca735f32c81509",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reservation_id"
                                ],
                                "properties": {
                                    "reservation_id": {
                                        "type": "string",
                                        "example": "uuid-reservation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Locação iniciada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/rentals/{rental}/checkin": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Realizar check-in (Devolver veículo)",
                "operationId": "9b55f6f93bb5ac36d281b12a38b3d20e",
                "parameters": [
                    {
                        "name": "rental",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locação finalizada e taxas calculadas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Listar minhas locações",
                "operationId": "c0ca7b519e22f654d2044c5a9c752342",
                "responses": {
                    "200": {
                        "description": "Minhas locações"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Listar minhas reservas",
                "operationId": "c0ead7ee8da750a341eeff509f3af8b9",
                "responses": {
                    "200": {
                        "description": "Minhas reservas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Criar uma reserva",
                "operationId": "ab4d7952bb3ef1ccd7724e8d862b034b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "string",
                                        "example": "uuid-vehicle"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva criada"
                    },
                    "422": {
                        "description": "Erro de disponibilidade"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/reservations/{reservation}/cancel": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancelar uma reserva",
                "operationId": "4b1925267f47b2cf5cd7325bd3419640",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Listar todos os veículos",
                "operationId": "f627fb85543e356532922246d50e6e2a",
                "responses": {
                    "200": {
                        "description": "Lista de veículos paginada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vehicles/available": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Listar veículos disponíveis (Cache)",
                "operationId": "04828ad8f4e4d9da0b3c3a60ba58f647",
                "responses": {
                    "200": {
                        "description": "Lista de veículos disponíveis"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Obter detalhes de um veículo",
                "operationId": "00b76a563ae78ef08e92480a3cd58a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do veículo"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}